extract common util types
