+++ packages/components/src/components/flow-item/flow-item.e2e.ts
@@ -12,11 +12,12 @@ import {
   themed,
 } from "../../tests/commonTests";
 import { html } from "../../../support/formatting";
-import { findAll, GlobalTestProps } from "../../tests/utils/puppeteer";
+import { findAll } from "../../tests/utils/puppeteer";
 import { scrollingContentHtml, scrollingHeightStyle } from "../panel/panel.e2e";
 import { IDS as PanelIDS } from "../panel/resources";
 import type { Action } from "../action/action";
 import { mockConsole } from "../../tests/utils/logging";
+import { GlobalTestProps } from "../../tests/utils/interfaces";
 import { CSS, SLOTS } from "./resources";
 import type { FlowItem } from "./flow-item";
