+++ packages/components/src/tests/commonTests/accessible.ts
@@ -1,7 +1,8 @@
 import axe from "axe-core";
 import { toHaveNoViolations } from "jest-axe";
 import { expect, it } from "vitest";
-import { GlobalTestProps, skipAnimations } from "../utils/puppeteer";
+import { skipAnimations } from "../utils/puppeteer";
+import { GlobalTestProps } from "../utils/interfaces";
 import { getTagAndPage } from "./puppeteer/utils";
 import { ComponentTestSetup, ComponentTag } from "./interfaces";
