+++ packages/components/src/utils/logger.browser.spec.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, Mock, vi } from "vitest";
 import { SetOptional } from "type-fest";
-import { GlobalTestProps } from "../tests/utils/puppeteer";
+import { GlobalTestProps } from "../tests/utils/interfaces";
 import { mockConsole } from "../tests/utils/logging";
 import { type LogLevel, loggedDeprecations, logger } from "./logger";
 import { type CalciteConfig, clearConfig } from "./config";
